Move commonly used code for web requests to common module
