+++ packages/azure-services/src/index.ts
@@ -23,5 +23,5 @@ export { BatchConfig } from './azure-batch/batch-config';
 export * from './azure-batch/pool-load-generator';
 export * from './azure-batch/job-task';
 export { BatchServiceClientProvider } from './ioc-types';
-export { ApplicationInsightsClient, ResponseWithBodyType } from './app-insights-api-client/application-insights-client';
+export { ApplicationInsightsClient } from './app-insights-api-client/application-insights-client';
 export * from './app-insights-api-client/query-response';
