+++ packages/common/src/web-requests/response-with-body-type.ts
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+import { Response } from 'got';
+
+// eslint-disable-next-line @typescript-eslint/ban-types
+export interface ResponseWithBodyType<T = {}> extends Response {
+    body: T;
+}
