+++ packages/common/src/web-requests/serializable-response.ts
@@ -1,7 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { ResponseWithBodyType } from './a11y-service-client';
+import { isNil } from 'lodash';
+import { ResponseWithBodyType } from './response-with-body-type';
 
 interface SerializableRequest {
     uri: URL;
@@ -13,22 +14,26 @@ export interface SerializableResponse<T = unknown> {
     statusCode: number;
     body: T;
     headers: { [key: string]: unknown };
-    request: SerializableRequest;
+    request?: SerializableRequest;
 }
 
 export type ResponseSerializer = typeof getSerializableResponse;
 
 export function getSerializableResponse<T>(response: ResponseWithBodyType<T>): SerializableResponse<T> {
-    const requestOptions = response.request?.options;
-
-    return {
+    const serializable: SerializableResponse<T> = {
         statusCode: response.statusCode,
         body: response.body,
         headers: response.headers,
-        request: {
-            uri: requestOptions?.url,
-            method: requestOptions?.method,
-            headers: requestOptions?.headers,
-        },
     };
+
+    if (!isNil(response.request)) {
+        const requestOptions = response.request.options;
+        serializable.request = {
+            uri: requestOptions.url,
+            method: requestOptions.method,
+            headers: requestOptions.headers,
+        };
+    }
+
+    return serializable;
 }
