+++ packages/web-api-client/src/index.ts
@@ -1,6 +1,5 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-export { A11yServiceClient, ResponseWithBodyType } from './a11y-service-client';
+export { A11yServiceClient } from './a11y-service-client';
 export { A11yServiceCredential } from './a11y-service-credential';
 export { A11yServiceClientProvider, a11yServiceClientTypeNames } from './a11y-service-client-types';
-export { SerializableResponse, getSerializableResponse, ResponseSerializer } from './serializable-response';
